refactor: Test 도메인 리팩토링
빌더가 아닌 생성자를 사용하여 엔티티 생성, 디렉토리 이름 변경
