+++ src/main/java/kr/co/fastcampus/yanabada/YanabadaApplication.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.Yanabada;
+package kr.co.fastcampus.yanabada;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
