+++ src/main/java/kr/co/fastcampus/yanabada/domain/test/entity/TestEntity.java
@@ -1,13 +1,12 @@
-package kr.co.fastcampus.Yanabada.domain.test.entity;
+package kr.co.fastcampus.yanabada.domain.test.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import kr.co.fastcampus.fastcatch.common.baseentity.BaseEntity;
+import kr.co.fastcampus.yanabada.common.baseentity.BaseEntity;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -26,10 +25,13 @@ public class TestEntity extends BaseEntity {
     @Column(nullable = false, length = 100)
     private String email;
 
-    @Builder
     private TestEntity(String name, String email) {
         this.name = name;
         this.email = email;
+    }
+
+    public static TestEntity create(String name, String email) {
+        return new TestEntity(name, email);
     }
 
 }
