+++ src/main/java/kr/co/fastcampus/yanabada/domain/test/repository/TestRepository.java
@@ -1,6 +1,6 @@
-package kr.co.fastcampus.Yanabada.domain.test.repository;
+package kr.co.fastcampus.yanabada.domain.test.repository;
 
-import kr.co.fastcampus.Yanabada.domain.test.entity.TestEntity;
+import kr.co.fastcampus.yanabada.domain.test.entity.TestEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TestRepository extends CrudRepository<TestEntity, Long> {
