+++ src/test/java/kr/co/fastcampus/yanabada/YanabadaApplicationTests.java
@@ -1,4 +1,4 @@
-package kr.co.fastcampus.Yanabada;
+package kr.co.fastcampus.yanabada;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
